pyproject.toml
[project]
name = "b3-asv-inference"
version = "0.1.0"
description = "ASVSpoof2021 Baseline Inference System"
requires-python = ">=3.8,<3.11"
dependencies = [
    "torch>=1.6.0,<2.0.0",
    "torchaudio>=0.6.0",
    "numpy>=1.18.0,<1.24.0",
    "scipy>=1.4.0",
    "scikit-learn>=0.24.0",
    "soundfile>=0.10.0",
    "librosa>=0.8.0",
    "pandas>=1.2.0",
    "h5py>=2.10.0",
    "pyyaml>=5.4.0",
    "numba==0.48.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
